fix: resolve all 7 bugs and 1 hygiene issue from audit report
BUG-1 (P0): Update 3 stale module version assertions in nebula_core.nelua
  - shader_compose: v0.6_phase3.7 -> v0.7_phase4.1
  - pipeline_factory: v0.7_phase3.7 -> v0.8_phase4.1
  - app_factory: v0.6_phase3.12 -> v0.7_phase4.1

BUG-2 (P0): Add dynamic NebulaBuf{N} exemption to axiom_validator.lua
  - Add is_nebula_buf_type() helper matching ^NebulaBuf%d+$ pattern
  - Apply exemption in both base_fields and state_fields checks

BUG-3 (P1): Add missing nebula_create_uniform_buffer() to renderer.nelua
  - Required by Slug pipeline init method

BUG-4 (P0): Fix Storage Buffer binding visibility in pipeline_factory.lua
  - binding 1 (instances): Fragment-only -> Vertex | Fragment
  - Vertex shader reads instances[inst] so VERTEX visibility is required

BUG-5 (P1): Extend nebula_create_storage_buffer() signature with data param
  - Add data: pointer and change size: uint64 -> csize for consistency
  - Upload initial data via wgpuQueueWriteBuffer when data != nilptr

BUG-6 (P1): Fix uint16/uint32 type mismatch for NEBULA_SLUG_BAND_REFS
  - app_factory.lua: #uint32 -> #uint16 for _refs_sz calculation
  - shader_compose.lua: array<u32> -> array<u16> for slug_band_refs
  - shader_compose.lua: add u32() cast when reading slug_band_refs

BUG-7 (P1): Add missing require for liberation_sans_slug_metrics in text_runtime.nelua
  - NebulaSlugGlyph and related symbols are defined there

HYGIENE-1: Remove duplicate nebula_create_vertex_buffer() from renderer.nelua
  - Phase 3.2.2 version (out_buffer param) removed; Phase 3.3 version kept

All 43 Phase 4.2.1 smoke tests pass. form_demo compiles and runs successfully.

+++ src/derive/shader_compose.lua
@@ -429,7 +429,7 @@ struct SlugBandMeta {
 }
 @group(0) @binding(1) var<storage, read> slug_curves:     array<SlugCurveData>;
 @group(0) @binding(2) var<storage, read> slug_band_metas: array<SlugBandMeta>;
-@group(0) @binding(3) var<storage, read> slug_band_refs:  array<u32>;
+@group(0) @binding(3) var<storage, read> slug_band_refs:  array<u16>;  // BUG-6 fix: CPU side is uint16
 
 // ---- 顶点输入/输出 ----
 struct SlugVertexInput {
@@ -536,7 +536,7 @@ fn fs_main(in: SlugVertexOutput) -> @location(0) vec4<f32> {
   let hband_meta_idx = band_base + u32(band_iy);
   let hband = slug_band_metas[hband_meta_idx];
   for (var ci: u32 = 0u; ci < hband.count; ci = ci + 1u) {
-    let curve_idx = slug_band_refs[hband.offset + ci];
+    let curve_idx = u32(slug_band_refs[hband.offset + ci]);  // BUG-6 fix: cast u16 to u32
     let c = slug_curves[curve_idx];
     let p12 = vec4<f32>(c.p0x, c.p0y, c.p1x, c.p1y) - vec4<f32>(render_coord, render_coord);
     let p3  = vec2<f32>(c.p2x, c.p2y) - render_coord;
@@ -562,7 +562,7 @@ fn fs_main(in: SlugVertexOutput) -> @location(0) vec4<f32> {
   let vband_meta_idx = band_base + u32(band_max_y + 1) + u32(band_ix);
   let vband = slug_band_metas[vband_meta_idx];
   for (var ci: u32 = 0u; ci < vband.count; ci = ci + 1u) {
-    let curve_idx = slug_band_refs[vband.offset + ci];
+    let curve_idx = u32(slug_band_refs[vband.offset + ci]);  // BUG-6 fix: cast u16 to u32
     let c = slug_curves[curve_idx];
     let p12 = vec4<f32>(c.p0x, c.p0y, c.p1x, c.p1y) - vec4<f32>(render_coord, render_coord);
     let p3  = vec2<f32>(c.p2x, c.p2y) - render_coord;
